+++ src/main/java/org/wikilayer/network/model/MyWiki.kt
@@ -24,6 +24,8 @@ data class MyWiki(
     val id: Long,
     val title: String = "",
     @field:JsonProperty("url_path") val urlPath: String = "",
+    @field:JsonProperty("icon_url") val iconUrl: String? = null,
+    @field:JsonProperty("pages_tree") val pagesTree: Boolean = false,
     @field:JsonProperty("updated_at") val updatedAt: Instant? = null,
     val visibility: WikiVisibility? = null,
     val mine: Boolean = false,